@@ -1,42 +1,32 @@
 {
-  "name": "ridehub_fe",
-  "version": "1.0.0",
-  "main": "expo-router/entry",
+  "name": "adminui",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "lint": "eslint ."
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@react-native-community/datetimepicker": "^8.4.4",
-    "clsx": "^2.1.1",
-    "expo": "^53.0.22",
-    "expo-constants": "~17.1.6",
-    "expo-linking": "~7.1.5",
-    "expo-router": "^5.1.5",
-    "expo-status-bar": "~2.2.3",
-    "nativewind": "^4.1.23",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-native": "^0.79.5",
-    "react-native-otp-entry": "^1.8.5",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1",
-    "react-native-simple-radio-button": "^2.7.4",
-    "react-native-web": "^0.20.0",
-    "tailwind-merge": "^3.0.1",
-    "tailwindcss": "^3"
+    "@tailwindcss/vite": "^4.1.13",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-router-dom": "^7.9.1",
+    "tailwindcss": "^4.1.13"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@types/react": "~19.0.10",
-    "@types/react-native-simple-radio-button": "^2.7.7",
-    "eslint": "^9.0.0",
-    "eslint-config-expo": "~9.2.0",
-    "typescript": "~5.8.3"
-  },
-  "private": true
+    "@eslint/js": "^9.33.0",
+    "@types/react": "^19.1.10",
+    "@types/react-dom": "^19.1.7",
+    "@vitejs/plugin-react": "^5.0.0",
+    "eslint": "^9.33.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
+    "globals": "^16.3.0",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.39.1",
+    "vite": "^7.1.2"
+  }
 }
